requirements.txt
flask
mysql_connector-python